style: moved test ensembles
