+++ src/ensembles.cpp
@@ -1,37 +1,17 @@
+/**
+ * @file ensembles.cpp
+ * @author your name (you@domain.com)
+ * @brief 
+ * @version 1
+ * @date 2024-06-10
+ * 
+ * @copyright Copyright (c) 2024
+ * 
+ */
 #include "ensembles.hpp"
-
 #ifndef TEST_VERSION
 #include "Particle.h"
 #else
 #include "scheduler_test_system.hpp"
 #endif
 
-
-void SS_ensembleAInit(DeploymentSchedule_t* pDeployment)
-{
-    return;
-}
-void SS_ensembleAFunc(DeploymentSchedule_t* pDeployment)
-{
-    delay(400);
-}
-
-
-void SS_ensembleBInit(DeploymentSchedule_t* pDeployment)
-{
-    return;
-}
-void SS_ensembleBFunc(DeploymentSchedule_t* pDeployment)
-{
-    delay(200);
-
-}
-
-void SS_ensembleCInit(DeploymentSchedule_t* pDeployment)
-{
-    return;
-}
-void SS_ensembleCFunc(DeploymentSchedule_t* pDeployment)
-{
-    delay(600);
-}
