+++ tests/CMakeLists.txt
@@ -7,16 +7,17 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 add_compile_definitions(TEST_VERSION=1 GTESTING=1)
 
 add_subdirectory(../external/googletest ./binary_dir)
-include_directories(../external/googletest/include  ../src/ ../src/cli/ ../test)
+include_directories(../external/googletest/include  ../src/ ../src/cli/ ../tests)
 
 set(GTEST_SOURCE_FILES
     scheduler_test_system.cpp
+    test_ensembles.cpp
     ../src/scheduler.cpp
-    scheduler_functions.cpp
     gtest.cpp)
 
 set(TEST_SOURCE_FILES
     scheduler_test_system.cpp
+    test_ensembles.cpp
     ../src/scheduler.cpp
     ../src/ensembles.cpp
     scheduler_test.cpp
