--- tests/scheduler_functions.cpp
@@ -1,26 +1,0 @@
-#include "scheduler_functions.hpp"
-
-void SS_ensembleAInit(DeploymentSchedule_t* pDeployment)
-{
-    return;
-}
-void SS_ensembleAFunc(DeploymentSchedule_t* pDeployment)
-{
-    return;
-}
-void SS_ensembleBInit(DeploymentSchedule_t* pDeployment)
-{
-    return;
-}
-void SS_ensembleBFunc(DeploymentSchedule_t* pDeployment)
-{
-    return;
-}
-void SS_ensembleCInit(DeploymentSchedule_t* pDeployment)
-{
-    return;
-}
-void SS_ensembleCFunc(DeploymentSchedule_t* pDeployment)
-{
-    return;
-}
