+++ tests/scheduler_test_system.cpp
@@ -1,3 +1,4 @@
+/**@todo Add documentation*/
 #include "scheduler_test_system.hpp"
 #include "conio.hpp"
 #include <chrono>
@@ -6,14 +7,14 @@
 #include <string>
 #include <cstdio>
 
-
+/**@todo Add documentation*/
 bool TestLog::operator==(const TestLog& rhs)
 {
     return  (name == rhs.name) &&
         (start == rhs.start) &&
         (end == rhs.end);
 };
-
+/**@todo Add documentation*/
 int SF_OSAL_printf(const char* fmt, ...)
 {
     va_list vargs;
@@ -22,19 +23,22 @@ int SF_OSAL_printf(const char* fmt, ...)
     va_end(vargs);
     return nBytes;
 }
-
+/**@todo Add documentation*/
 uint32_t millis()
 {
     return testTime;
 }
+/**@todo Add documentation*/
 void addTime(uint32_t add)
 {
     testTime += add;
 }
+/**@todo Add documentation*/
 void setTime(uint32_t set)
 {
     testTime = set;
 }
+/**@todo Add documentation*/
 void delay(uint32_t time)
 {
     addTime(time);
