+++ tests/scheduler_test_system.hpp
@@ -1,15 +1,21 @@
+/** @todo Add documentation*/
 #ifndef __SCHEDULER__TEST__HPP_
 #define __SCHEDULER__TEST__HPP_
 #include <stdint.h>
 
+/** @todo Add documentation*/
 typedef uint32_t system_tick_t;
+/** @todo Add documentation*/
 static int testTime = 0;
+/** @todo Add documentation*/
 void addTime(uint32_t add);
-
+/** @todo Add documentation*/
 void setTime(uint32_t set);
+/** @todo Add documentation*/
 void delay(uint32_t add);
+/** @todo Add documentation*/
 uint32_t millis();
-
+/** @todo Add documentation*/
 struct TestLog
 {
     char name;
