+++ tests/test_ensembles.cpp
@@ -1,0 +1,46 @@
+/**
+ * @file ensembles.cpp
+ * @author your name (you@domain.com)
+ * @brief 
+ * @version 0.1
+ * @date 2024-06-10
+ * 
+ * @copyright Copyright (c) 2024
+ * 
+ */
+#include "test_ensembles.hpp"
+#include "scheduler_test_system.hpp"
+
+/**@todo Add documentation*/
+void SS_ensembleAInit(DeploymentSchedule_t* pDeployment)
+{
+    return;
+}
+/**@todo Add documentation*/
+void SS_ensembleAFunc(DeploymentSchedule_t* pDeployment)
+{
+    delay(400);
+}
+
+/**@todo Add documentation*/
+void SS_ensembleBInit(DeploymentSchedule_t* pDeployment)
+{
+    return;
+}
+/**@todo Add documentation*/
+void SS_ensembleBFunc(DeploymentSchedule_t* pDeployment)
+{
+    delay(200);
+
+}
+
+/**@todo Add documentation*/
+void SS_ensembleCInit(DeploymentSchedule_t* pDeployment)
+{
+    return;
+}
+/**@todo Add documentation*/
+void SS_ensembleCFunc(DeploymentSchedule_t* pDeployment)
+{
+    delay(600);
+}
