+++ tests/test_ensembles.hpp
@@ -1,0 +1,24 @@
+/**
+ * @file ensembles.hpp
+ * @brief Header file to declare various ensembles
+ */
+
+#ifndef __TEST_ENSEMBLES_HPP__
+#define __TEST_ENSEMBLES_HPP__
+
+#include "scheduler.hpp"
+/** @todo Add documentation*/
+void SS_ensembleAInit(DeploymentSchedule_t* pDeployment);
+/** @todo Add documentation*/
+void SS_ensembleAFunc(DeploymentSchedule_t* pDeployment);
+/** @todo Add documentation*/
+void SS_ensembleBInit(DeploymentSchedule_t* pDeployment);
+/** @todo Add documentation*/
+void SS_ensembleBFunc(DeploymentSchedule_t* pDeployment);
+/** @todo Add documentation*/
+void SS_ensembleCInit(DeploymentSchedule_t* pDeployment);
+/** @todo Add documentation*/
+void SS_ensembleCFunc(DeploymentSchedule_t* pDeployment);
+
+
+#endif //__TEST_ENSEMBLES_HPP__
